refactor(backend): move provider passport package to root
